new manthana radiology/manthana-backend/services/21_cxr_medgemma/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
Pillow==10.4.0
openai==1.40.0
httpx>=0.28.0

# MedGemma (shared/medical_document_parser) — same stack as lab_report
torch>=2.7.0
transformers>=4.52.0
accelerate>=1.0.0
huggingface_hub>=0.24.0
modelscope>=1.9.0
